Renames and docs fixes
